requirements.txt
certifi==2022.5.18.1
future==0.18.2
numpy==1.22.3
Pillow==9.0.1
pyglet==1.5.27
six==1.16.0
wincertstore==0.2

mkl-random==1.2.2
mkl-service==2.4.0
mkl-fft==1.3.1